--- src/main/AndroidManifest.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.bpsl">
-
-    <application
-        android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
-        android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-        <activity
-            android:name=".MainActivity"
-            android:label="@string/app_name"
-            android:theme="@style/AppTheme.NoActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
-            android:name=".NewListDescription"
-            android:label="@string/app_name">
-        </activity>
-        <activity
-            android:name=".NewListSearch"
-            android:label="@string/app_name">
-        </activity>
-        <activity
-            android:name=".NewListOption"
-            android:label="@string/app_name">
-        </activity>
-        <activity
-            android:name=".NewListSave"
-            android:label="@string/app_name">
-        </activity>
-        <activity
-            android:name=".SavedList"
-            android:label="@string/app_name">
-        </activity>
-    </application>
-
-</manifest>
