--- src/main/java/com/example/bpsl/MainActivity.java
@@ -1,76 +1,0 @@
-package com.example.bpsl;
-
-import android.annotation.SuppressLint;
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.widget.Button;
-import android.widget.CursorAdapter;
-import android.widget.ListView;
-
-public class MainActivity extends AppCompatActivity {
-
-    private Button MainButton;
-
-//    private CursorAdapter cursorAdapter;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        MainButton = (Button) findViewById(R.id.newlistbutton);
-        MainButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-//                setContentView(R.layout.newlistdescription);
-                Intent intent = new Intent(v.getContext(), NewListDescription.class);
-                startActivity(intent);
-
-            }
-        });
-
-        MainButton = (Button) findViewById(R.id.savedbutton);
-        MainButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-//                setContentView(R.layout.savelist);
-                Intent intent = new Intent(v.getContext(), SavedList.class);
-                startActivity(intent);
-            }
-        });
-
-
-
-    }
-
-
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_settings) {
-            return true;
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
-}
