--- src/main/java/com/example/bpsl/NewListOption.java
@@ -1,63 +1,0 @@
-package com.example.bpsl;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.view.Menu;
-import android.view.View;
-import android.widget.Button;
-
-public class NewListOption extends AppCompatActivity {
-
-    private Button buttonOne;
-    private Button buttonTwo;
-    private Button buttonThree;
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.newlistoption);
-
-        buttonOne = (Button) findViewById(R.id.buttonOptionOne);
-        buttonOne.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(NewListOption.this, NewListSave.class);
-                startActivity(intent);
-            }
-        });
-
-        buttonTwo = (Button) findViewById(R.id.buttonOptionTwo);
-        buttonTwo.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(NewListOption.this, NewListSave.class);
-                startActivity(intent);
-            }
-        });
-
-        buttonThree = (Button) findViewById(R.id.buttonOptionThree);
-        buttonThree.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(v.getContext(), NewListSave.class);
-                startActivity(intent);
-            }
-        });
-    }
-
-//    public void openNewListSave(View v){
-//        Intent intent = new Intent(v.getContext(), NewListSave.class);
-//        startActivity(intent);
-//
-//    }
-
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main, menu);
-        return true;
-    }
-}
