--- src/main/res/layout/newlistoption.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <TextView
-        android:id="@+id/textOption"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="164dp"
-        android:layout_marginLeft="164dp"
-        android:layout_marginTop="55dp"
-        android:text="Option"
-        android:textSize="25dp"
-        tools:layout_editor_absoluteX="155dp"
-        tools:layout_editor_absoluteY="41dp" />
-
-    <Button
-        android:id="@+id/buttonOptionOne"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="146dp"
-        android:layout_marginLeft="146dp"
-        android:layout_marginTop="115dp"
-        android:text="Option 1"
-        android:textSize="20dp"
-        android:clickable="true"
-        tools:layout_editor_absoluteX="142dp"
-        tools:layout_editor_absoluteY="118dp" />
-
-    <Button
-        android:id="@+id/buttonOptionTwo"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="147dp"
-        android:layout_marginLeft="147dp"
-        android:layout_marginTop="205dp"
-        android:text="Option 2"
-        android:textSize="20dp"
-        android:clickable="true"
-        tools:layout_editor_absoluteX="142dp"
-        tools:layout_editor_absoluteY="256dp" />
-
-    <Button
-        android:id="@+id/buttonOptionThree"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="149dp"
-        android:layout_marginLeft="149dp"
-        android:layout_marginTop="294dp"
-        android:text="Option 3"
-        android:textSize="20dp"
-        android:clickable="true"
-        tools:layout_editor_absoluteX="142dp"
-        tools:layout_editor_absoluteY="393dp" />
-
-</RelativeLayout>
